requirements.txt
avalanche_lib==0.5.0
mlflow==2.10.2
numpy==1.24.1
Pillow==11.0.0
scikit_learn==1.4.2
torch==2.2.1+cu118
torchvision==0.17.1+cu118
